pyproject.toml
[project]
name = "dokploy-api"
version = "0.1.0"
authors = [
    { name = "pythoninthegrass", email = "4097471+pythoninthegrass@users.noreply.github.com" }
]

requires-python = ">=3.13,<3.14"

dependencies = [
    "requests>=2.32.5",
    "click>=8.1.0",
    "pyyaml>=6.0",
    "rich>=13.0.0",
    "python-decouple>=3.8",
]

[project.scripts]
dokploy = "api:cli"

[project.optional-dependencies]
dev = [
    "ruff>=0.14.4",
]

test = [
    "coverage<8.0.0,>=7.6.1",
    "hypothesis[cli]<7.0.0,>=6.112.1",
    "pytest<9.0.0,>=8.3.3",
    "pytest-asyncio<1.0.0,>=0.24.0",
    "pytest-cov>=6.1.1",
    "pytest-datafiles<4.0.0,>=3.0.0",
    "pytest-xdist<4.0.0,>=3.6.1",
    "responses>=0.25.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
python_classes = "Test*"
asyncio_mode = "auto"
markers = [
    "unit: marks tests as unit tests",
    "integration: marks tests as integration tests",
    "e2e: marks tests as end-to-end tests",
    "benchmark: marks performance benchmark tests"
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = ["api", "extract", "schema", "validate"]
packages = ["lib"]
